Moved events from os.hpp to events.hpp
