+++ include/physics-system.hpp
@@ -10,6 +10,7 @@
 #include "entity.hpp"
 #include "multiton.hpp"
 #include "command-queue.hpp"
+#include "event-system.hpp"
 
 namespace tec {
 	struct CollisionBody;
