+++ src/main.cpp
@@ -1,4 +1,5 @@
 #include "os.hpp"
+#include "events.hpp"
 #include "reflection.hpp"
 #include "render-system.hpp"
 #include "physics-system.hpp"
