+++ src/physics-system.cpp
@@ -1,9 +1,10 @@
 #include "physics-system.hpp"
 
 #include "components/collisionbody.hpp"
-#include "entity.hpp"
 #include "components/transforms.hpp"
 #include "component-update-system.hpp"
+#include "entity.hpp"
+#include "events.hpp"
 
 #include "physics/physics-debug-drawer.hpp"
 #include <BulletCollision/Gimpact/btGImpactShape.h>
